kind of silly for this task to be so specific about formmating of an error, this is as far as im taking this one

+++ 0x03-asynchronous/4-get_sum.dart
@@ -1,7 +1,7 @@
 import '4-util.dart';
 import 'dart:convert';
 
-Future<int> calculateTotal() async {
+Future<double> calculateTotal() async {
   try {
     var user = await fetchUserData();
     var userId = jsonDecode(user)['id'];
